requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
Flask-Login
psycopg2-binary
python-dotenv
python-telegram-bot
requests
gunicorn

instaloader

